+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to `parsenic` will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to
 [Semantic Versioning].
 
+## [0.2.0] - Unreleased
+### Added
+ - `error` module with error types
+ - `result` module with `Result` type aliases
+
+### Changed
+ - Bounds of `Int` and `UInt`
+ - Now depends on `traitful` for extension traits
+
+### Fixed
+ - Broken links in README and docs
+ - Parsing `u8()` not advancing reader
+
 ## [0.1.0] - 2023-07-16
 ### Added
  - `be` module